Create README.md
The table will contain a list of developers. Each row contains a name, an age, and programming language. When a user types into the input field, the table will filter and return the rows which match the value entered.

During the process of building this, you will learn about some CSS properties as well as DOM access/manipulation with JavaScript. With just HTML, CSS, and JavaScript you can create some really stylish elements with advanced features.